Adding ESP8266 Hardware SPI
connect the DotStar to Pin 13 (Data) and Pin 14 (Clock) of an ESP8266 Huzzah Module to use its SPI.

The bitbang approach is ~3 times slower for me (300 Pixels):

    Adafruit_DotStar strip = Adafruit_DotStar(300, 13, 14, DOTSTAR_BRG);
    // strip.show => 14000µs

    Adafruit_DotStar strip = Adafruit_DotStar(300, DOTSTAR_BRG);
    // strip.show => 4000µs

I also experimented with faster Clock Dividers, but for the amount of pixels i'm pushing, a faster clock seems to be a problem. The Pixels started acting weird after 100-150 pixels.